Move in_play and play_filter to utils
