+++ torchbox/templatetags/torchbox_tags.py
@@ -1,9 +1,8 @@
-from itertools import chain, cycle, islice
-
 from django import template
 from django.conf import settings
 
 from torchbox.models import *
+from torchbox.utils import *
 
 register = template.Library()
 
@@ -66,21 +65,7 @@ def content_type(value):
 
 @register.filter
 def in_play(page):
-    """
-    Check to see if a page is in the Play section. A page is in the Play
-    section if it has 'show_in_play_menu' set to True, or one of its
-    ancestors does.
-    """
-    if not page:
-        return False
-
-    if getattr(page.specific, 'show_in_play_menu', False):
-        return True
-
-    return any(
-        getattr(ancestor.specific, 'show_in_play_menu', False)
-        for ancestor in page.get_ancestors()
-    )
+    return is_in_play(page)
 
 
 @register.inclusion_tag('torchbox/tags/top_menu.html', takes_context=True)
@@ -275,32 +260,3 @@ def time_display(time):
 
     # Join and return
     return "".join([hour_string, minute_string, pm_string])
-
-
-def play_filter(pages, number):
-    """
-    Given an iterable of Pages, return a specified number that
-    are not in the Play section.
-    """
-    result = []
-    for page in pages:
-        if len(result) > (number - 1):
-            break
-        if not in_play(page):
-            result.append(page)
-    return result
-
-
-# https://docs.python.org/2/library/itertools.html#recipes
-def roundrobin(*iterables):
-    "roundrobin('ABC', 'D', 'EF') --> A D E B F C"
-    # Recipe credited to George Sakkis
-    pending = len(iterables)
-    nexts = cycle(iter(it).next for it in iterables)
-    while pending:
-        try:
-            for next in nexts:
-                yield next()
-        except StopIteration:
-            pending -= 1
-            nexts = cycle(islice(nexts, pending))
